+++ ubelt/__init__.py
@@ -22,7 +22,7 @@ Testing:
     xdoctest ubelt
 """
 
-__version__ = '0.9.0'
+__version__ = '0.9.1'
 
 __submodules__ = [
     'util_arg',